requirements.txt
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
debtcollector==3.0.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
httplib2==0.22.0
idna==3.10
mysql-connector-python==9.1.0
netaddr==1.3.0
pytest>=7.0.0
oauthlib==3.2.2
oauth2client==4.1.3
oslo.config==9.6.0
oslo.i18n==6.4.0
pbr==6.1.0
protobuf==5.28.3
python-dateutil==2.9.0.post0
pytz==2022.7.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
soupsieve==2.6
stevedore==5.3.0
uritemplate==4.1.1
urllib3==2.2.3
wrapt==1.16.0
